+++ src/main/java/com/authlete/jaxrs/api/AuthleteApiImpl.java
@@ -87,6 +87,8 @@ import com.authlete.common.dto.TokenIssueResponse;
 import com.authlete.common.dto.TokenListResponse;
 import com.authlete.common.dto.TokenRequest;
 import com.authlete.common.dto.TokenResponse;
+import com.authlete.common.dto.TokenRevokeRequest;
+import com.authlete.common.dto.TokenRevokeResponse;
 import com.authlete.common.dto.TokenUpdateRequest;
 import com.authlete.common.dto.TokenUpdateResponse;
 import com.authlete.common.dto.UserInfoIssueRequest;
@@ -114,6 +116,7 @@ public class AuthleteApiImpl extends AuthleteApiJaxrsImpl
     private static final String AUTH_TOKEN_FAIL_API_PATH                      = "/api/auth/token/fail";
     private static final String AUTH_TOKEN_GET_LIST_API_PATH                  = "/api/auth/token/get/list";
     private static final String AUTH_TOKEN_ISSUE_API_PATH                     = "/api/auth/token/issue";
+    private static final String AUTH_TOKEN_REVOKE_API_PATH                    = "/api/auth/token/revoke";
     private static final String AUTH_TOKEN_UPDATE_API_PATH                    = "/api/auth/token/update";
     private static final String AUTH_REVOCATION_API_PATH                      = "/api/auth/revocation";
     private static final String AUTH_USERINFO_API_PATH                        = "/api/auth/userinfo";
@@ -542,6 +545,18 @@ public class AuthleteApiImpl extends AuthleteApiJaxrsImpl
         return executeApiCall(
                 new ServicePostApiCaller<TokenIssueResponse>(
                         TokenIssueResponse.class, request, AUTH_TOKEN_ISSUE_API_PATH));
+    }
+
+
+    /**
+     * Call {@code /api/auth/token/revoke} API.
+     */
+    @Override
+    public TokenRevokeResponse tokenRevoke(TokenRevokeRequest request) throws AuthleteApiException
+    {
+        return executeApiCall(
+                new ServicePostApiCaller<TokenRevokeResponse>(
+                        TokenRevokeResponse.class, request, AUTH_TOKEN_REVOKE_API_PATH));
     }